requirements.txt
behave==1.2.6
certifi==2018.11.29
chardet==3.0.4
docker==3.6.0
docker-pycreds==0.4.0
idna==2.8
paho-mqtt==1.4.0
parse==1.9.0
parse-type==0.4.2
requests==2.21.0
six==1.12.0
urllib3==1.24.1
websocket-client==0.54.0
